refactor(git): split repository in multiple modules
