+++ src/git/hook.rs
@@ -7,8 +7,8 @@ use crate::CocoGitto;
 
 use anyhow::{anyhow, Result};
 
-pub static PRE_PUSH_HOOK: &[u8] = include_bytes!("assets/pre-push");
-pub static PREPARE_COMMIT_HOOK: &[u8] = include_bytes!("assets/pre-commit");
+pub(crate) static PRE_PUSH_HOOK: &[u8] = include_bytes!("assets/pre-push");
+pub(crate) static PREPARE_COMMIT_HOOK: &[u8] = include_bytes!("assets/pre-commit");
 const PRE_COMMIT_HOOK_PATH: &str = ".git/hooks/pre-commit";
 const PRE_PUSH_HOOK_PATH: &str = ".git/hooks/pre-push";
